apai-landing/requirements.txt
flask
google-generativeai
matplotlib
scipy
numpy
psycopg2-binary